requirements.txt
alembic==0.9.3
aniso8601==1.2.1
bcrypt==3.1.3
cffi==1.10.0
click==6.7
Flask==0.12.2
Flask-API==0.7.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-Migrate==2.0.4
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.5
nose==1.3.7
psycopg2==2.7.3
pycparser==2.18
PyJWT==1.4.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.12
Werkzeug==0.12.2
